Events.delay
Brand new method just for this feature (and maybe more in future).
When an event has a future outcome it can rely on this method. It will
a. execute the "result" callback after <timeout> seconds
b. store a state name related to said event in State as "wait": {"eventName": "residualTimeout"}
c. if invoked on init() or similar onload function call along with the required name and callback, the latter will be executed when residual time expires