Unified action load on event and listener
This changes quite a lot from existing file for Mysterious Wanderers.
Actions for wood wanderers (short and large) and fur wanderers (short and large) are now unified (reward and probability are parametric). This makes it easier to call the action on event and to initialize the listener.
Instead of programming a complex init() for this, I chose a get() instance. This allows the "delayed" property to rely upon actions (defined as event methods) while Event.Room is initialised (it would throw an "undefined" error otherwise).
Timeouts have been moved to Event.delay() method.