destroyHuts
New method for Outside. It extracts a random number for the hut to be destroyed, so that it can also destroy the less populated one.
It takes 2 parameters. The first is the number of huts to be destroyed (hehehe), and the second is a boolean. If true, it can destroy empty huts as well. It is not set on hut fire event, as it's unlikely that an empty hut takes fire spontaneously. But if you had in mind to include, IDK, earthquakes, plane bombing or similar, they're things which do not distinguish between wood and flesh.
Also, it returns the number of victims. So you can <s>know how sadistic the script is</s> trigger different actions, depending on the death rate.