Tab Navigation
Apart from calls inside World module, now it's all here.
Engine.tabNavigations is true and allows navigating tabs.
When moving to World, it is set to false. The 2 actions that exit World (dying and coming back) trigger Engine.restoreNavigation after some milliseconds, necessary to prevent the navigation to Ship or Outside when the last step was towards right or left.