Events adapted to Delay stuff
This is the most obvious change.
The 4 MW events have now a structure compatible with Delay methods. An actions (which is also called by recallDelay if needed) triggers reward and notification, and it takes an optional time delay in seconds (if not provided, actions take it from state).
Onload methods invoke the actions associating them with a probability and a defined time (60 secons as it was before).
Note: I changed scene names from 100wood to wood100 (and same for the other ones), as a name starting with a number gives some issues with retrieving the objects using the full path.