Room and the Compass
Added new property for room. It has a boolean value reflecting the fact a player has a compass or not when the game is loaded.
If not, when the compass is added to the box, it triggers openPath.
This prevents triggering Path.openPath when reloading and solves every issue related to function calls timing (currently, when the compass is bought from the Nomad, openPath is triggered before stores.compass is updated; when bought normally World.dir is available some milliseconds too late due to the elaboration of the Map, so the inconsistency makes difficult defining the call stack. Now openPath is invariably called when the Compass is inserted for the first time; therefore the Compass row is always available when World.dir is generated and the tooltip can be appended)