Textarea autoselection
As textarea is used in 2 cases throughout the game, i.e. on save import and export, and in both cases it's useful to have an autofocus/autoselect state on it, here it is. Save textarea is autoselected, and therefore on import (where there's nothing to select) it is focused and ready for pasting.