No cooldown on reload
Let's ease the life of players.
Embark button kept cooling down during exploration, and it's saved. So if a player reloaded s/he found out residual cooldown for it.
Now Embark cooldown is reset when starting exploration (and in case, it is restarted on death).
Embark reset when returning home is therefore removed.