Changed size of languages menu
The Portuguese (brazil) option was too long and wrapping around to the next line, appearing right over the next language in the menu. Increased the width to accommodate in one line. Also increased the max height as the last language option wasn't visible. The menu can take one more language before needing to increase the height.