Removed background-color for config lists.
Background colors for unordered list in options were set to match body.
This created a problem during and after the space game when the body and
text were inverted, but the ul was not. See Issue #432.

Tried this via Chrome's dev tools while at the effected screen in dark
theme, and it seemed to work.