Fix spelling mistake
For each location where the string "another beast, draw by the noise, leaps out of a copse of trees." is located, "draw" is replaced with "drawn".  This is changed in all strings.po for all languages, as well as setpieces.js and adarkroom.pot.

This is because "drawn" is the past participle of "draw".  The phrase "drawn by the noise" functions as an adjectival phrase modifying the noun "beast." It indicates that the beast is attracted or compelled to leap out of the copse of trees due to the noise, hence requiring the past participle.